DT-1023  Removes result.json and winnowing.wfp files before resuming scan

+++ src/main/task/scanner/rescan/RescanTask.ts
@@ -1,4 +1,5 @@
  import log from "electron-log";
+import fs from 'fs';
 import { ScannerStage, ScanState } from "../../../../api/types";
 import { BaseScannerTask } from "../BaseScannerTask";
 import { Scanner } from "../types";
@@ -7,8 +8,8 @@ import { licenseService } from "../../../services/LicenseService";
 import { rescanService } from "../../../services/RescanService";
 import { IDispatch } from "../dispatcher/IDispatch";
 import { IScannerInputAdapter } from "../adapter/IScannerInputAdapter";
- import fs from 'fs';
- import { treeService } from '../../../services/TreeService';
+ import { fileExists } from '../../../utils/utils';
+
 
 export abstract class RescanTask<TDispatcher extends IDispatch,TInputScannerAdapter extends IScannerInputAdapter> extends BaseScannerTask<TDispatcher,TInputScannerAdapter> {
 
@@ -23,8 +24,8 @@ export abstract class RescanTask<TDispatcher extends IDispatch,TInputScannerAdap
   public abstract reScan(): Promise<void>;
 
   public async init(): Promise<void> {
-    await fs.promises.unlink(`${this.project.getMyPath()}/result.json`);
-    await fs.promises.unlink(`${this.project.getMyPath()}/winnowing.wfp`);
+    if (await fileExists(`${this.project.getMyPath()}/result.json`)) await fs.promises.unlink(`${this.project.getMyPath()}/result.json`);
+    if (await fileExists(`${this.project.getMyPath()}/winnowing.wfp`)) await fs.promises.unlink(`${this.project.getMyPath()}/winnowing.wfp`);
     await super.init();
   }
